+++ Assets/Scripts/Serialize/SerializerObject/SerializerObject.cs
@@ -86,13 +86,13 @@ namespace R1Engine
         public abstract Pointer SerializePointer(Pointer obj, Pointer anchor = null, bool allowInvalid = false, string name = null);
         public abstract Pointer<T> SerializePointer<T>(Pointer<T> obj, Pointer anchor = null, bool resolve = false, Action<T> onPreSerialize = null, bool allowInvalid = false, string name = null) where T : R1Serializable, new();
 
-        public abstract T[] SerializeArray<T>(T[] obj, decimal count, string name = null);
-        public abstract T[] SerializeObjectArray<T>(T[] obj, decimal count, Action<T> onPreSerialize = null, string name = null) where T : R1Serializable, new();
-        public abstract Pointer[] SerializePointerArray(Pointer[] obj, decimal count, Pointer anchor = null, bool allowInvalid = false, string name = null);
-        public abstract Pointer<T>[] SerializePointerArray<T>(Pointer<T>[] obj, decimal count, Pointer anchor = null, bool resolve = false, Action<T> onPreSerialize = null, bool allowInvalid = false, string name = null) where T : R1Serializable, new();
+        public abstract T[] SerializeArray<T>(T[] obj, long count, string name = null);
+        public abstract T[] SerializeObjectArray<T>(T[] obj, long count, Action<T> onPreSerialize = null, string name = null) where T : R1Serializable, new();
+        public abstract Pointer[] SerializePointerArray(Pointer[] obj, long count, Pointer anchor = null, bool allowInvalid = false, string name = null);
+        public abstract Pointer<T>[] SerializePointerArray<T>(Pointer<T>[] obj, long count, Pointer anchor = null, bool resolve = false, Action<T> onPreSerialize = null, bool allowInvalid = false, string name = null) where T : R1Serializable, new();
 
-        public abstract string SerializeString(string obj, decimal? length = null, Encoding encoding = null, string name = null);
-        public abstract string[] SerializeStringArray(string[] obj, decimal count, int length, Encoding encoding = null, string name = null);
+        public abstract string SerializeString(string obj, long? length = null, Encoding encoding = null, string name = null);
+        public abstract string[] SerializeStringArray(string[] obj, long count, int length, Encoding encoding = null, string name = null);
 
         public abstract T[] SerializeArraySize<T, U>(T[] obj, string name = null) where U : struct;
